# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminDashboardResponse": {
        "additionalProperties": true,
        "properties": {
          "failed_message_count": {
            "title": "Failed Message Count",
            "type": "integer"
          },
          "failed_messages": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "title": "Failed Messages",
            "type": "array"
          },
          "high_risk_event_count": {
            "title": "High Risk Event Count",
            "type": "integer"
          },
          "high_risk_events": {
            "items": {
              "$ref": "#/components/schemas/HighRiskEventResponse"
            },
            "title": "High Risk Events",
            "type": "array"
          },
          "integration_health": {
            "$ref": "#/components/schemas/IntegrationHealthResponse"
          },
          "pending_count": {
            "title": "Pending Count",
            "type": "integer"
          },
          "pending_items": {
            "items": {
              "$ref": "#/components/schemas/PendingHandoverResponse"
            },
            "title": "Pending Items",
            "type": "array"
          }
        },
        "required": [
          "pending_count",
          "failed_message_count",
          "high_risk_event_count",
          "pending_items",
          "failed_messages",
          "high_risk_events",
          "integration_health"
        ],
        "title": "AdminDashboardResponse",
        "type": "object"
      },
...
      "Envelope_AdminDashboardResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminDashboardResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[AdminDashboardResponse]",
        "type": "object"
      },
...
      "Envelope_MarkAllReadResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarkAllReadResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[MarkAllReadResponse]",
        "type": "object"
      },
...
      "Envelope_MessageListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[MessageListResponse]",
        "type": "object"
      },
      "Envelope_MessageResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[MessageResponse]",
        "type": "object"
      },
...
      "Envelope_list_StaffOptionResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StaffOptionResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[list[StaffOptionResponse]]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HighRiskEventResponse": {
        "additionalProperties": true,
        "properties": {
          "business_no": {
            "title": "Business No",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "business_no",
          "location",
          "risk_level",
          "status",
          "updated_at"
        ],
        "title": "HighRiskEventResponse",
        "type": "object"
      },
...
      "IntegrationHealthResponse": {
        "additionalProperties": true,
        "properties": {
          "database": {
            "title": "Database",
            "type": "string"
          },
          "message_delivery": {
            "title": "Message Delivery",
            "type": "string"
          },
          "model_gateway": {
            "title": "Model Gateway",
            "type": "string"
          }
        },
        "required": [
          "database",
          "message_delivery",
          "model_gateway"
        ],
        "title": "IntegrationHealthResponse",
        "type": "object"
      },
...
      "MarkAllReadResponse": {
        "additionalProperties": true,
        "properties": {
          "read_at": {
            "title": "Read At",
            "type": "string"
          },
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "updated_count",
          "read_at"
        ],
        "title": "MarkAllReadResponse",
        "type": "object"
      },
...
      "MessageListResponse": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "MessageListResponse",
        "type": "object"
      },
      "MessageResponse": {
        "additionalProperties": true,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "business_type": {
            "title": "Business Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "fallback_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Contact"
          },
          "handover_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handover Status"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_read": {
            "title": "Is Read",
            "type": "boolean"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "max_retry_count": {
            "title": "Max Retry Count",
            "type": "integer"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "retry_exhausted": {
            "title": "Retry Exhausted",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "business_type",
          "resource_id",
          "title",
          "body",
          "status",
          "is_read",
          "read_at",
          "retry_count",
          "max_retry_count",
          "retry_exhausted",
          "last_error",
          "handover_status",
          "fallback_contact",
          "created_at",
          "updated_at"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "PendingHandoverResponse": {
        "additionalProperties": true,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "queue": {
            "title": "Queue",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "queue",
          "summary",
          "status",
          "created_at"
        ],
        "title": "PendingHandoverResponse",
        "type": "object"
      },
...
      "StaffOptionResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "role"
        ],
        "title": "StaffOptionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "物业社区管理智能体 — 统一后端服务",
    "title": "Property Community Management Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/dashboard": {
      "get": {
        "operationId": "admin_dashboard_api_admin_dashboard_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminDashboardResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Dashboard",
        "tags": [
          "platform-operations"
        ]
      }
    },
...
    "/api/messages": {
      "get": {
        "operationId": "list_messages_api_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "SENT",
                    "FAILED",
                    "READ",
                    "UNREAD"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "business_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "REPAIR",
                    "ANNOUNCEMENT",
                    "BILLING",
                    "INSPECTION"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MessageListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "platform-operations"
        ]
      }
    },
    "/api/messages/read-all": {
      "post": {
        "operationId": "mark_all_messages_read_api_messages_read_all_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MarkAllReadResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark All Messages Read",
        "tags": [
          "platform-operations"
        ]
      }
    },
    "/api/messages/{message_id}/read": {
      "post": {
        "operationId": "mark_message_read_api_messages__message_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MessageResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Message Read",
        "tags": [
          "platform-operations"
        ]
      }
    },
...
    "/api/staff": {
      "get": {
        "description": "Return active, currently authorized staff from the caller's community.",
        "operationId": "list_staff_api_staff_get",
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "required": true,
            "schema": {
              "enum": [
                "REPAIR_WORKER",
                "SECURITY_GUARD"
              ],
              "title": "Role",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_StaffOptionResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Staff",
        "tags": [
          "platform-operations"
        ]
      }
    },
...
  }
}